Refactored getRegionInfo method
