+++ bigtable-hbase-1.x-parent/bigtable-hbase-1.x/src/main/java/com/google/cloud/bigtable/hbase1_x/BigtableConnection.java
@@ -19,9 +19,6 @@ import java.io.IOException;
 import java.util.concurrent.ExecutorService;
 
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.hbase.HRegionInfo;
-import org.apache.hadoop.hbase.HRegionLocation;
-import org.apache.hadoop.hbase.ServerName;
 import org.apache.hadoop.hbase.TableName;
 import org.apache.hadoop.hbase.client.AbstractBigtableConnection;
 import org.apache.hadoop.hbase.client.Admin;
@@ -29,7 +26,6 @@ import org.apache.hadoop.hbase.client.Table;
 import org.apache.hadoop.hbase.security.User;
 
 import com.google.cloud.bigtable.hbase.AbstractBigtableTable;
-import com.google.cloud.bigtable.hbase.adapters.SampledRowKeysAdapter;
 
 /**
  * HBase 1.x specific implementation of {@link AbstractBigtableConnection}.
@@ -59,18 +55,6 @@ public class BigtableConnection extends AbstractBigtableConnection {
   @Override
   public Admin getAdmin() throws IOException {
     return new BigtableAdmin(this);
-  }
-
-  @Override
-  protected SampledRowKeysAdapter createSampledRowKeysAdapter(TableName tableName,
-      ServerName serverName) {
-    return new SampledRowKeysAdapter(tableName, serverName) {
-      @Override
-      protected HRegionLocation createRegionLocation(byte[] startKey, byte[] endKey) {
-        HRegionInfo hRegionInfo = new HRegionInfo(tableName, startKey, endKey);
-        return new HRegionLocation(hRegionInfo, serverName);
-      }
-    };
   }
 
   @Override
