+++ bigtable-hbase-2.x-parent/bigtable-hbase-2.x/src/main/java/com/google/cloud/bigtable/hbase2_x/BigtableAsyncAdmin.java
@@ -26,7 +26,6 @@ import com.google.bigtable.admin.v2.ListTablesRequest;
 import com.google.bigtable.admin.v2.Snapshot;
 import com.google.bigtable.admin.v2.SnapshotTableRequest;
 import com.google.bigtable.admin.v2.Table;
-import com.google.bigtable.v2.SampleRowKeysRequest;
 import com.google.cloud.bigtable.config.BigtableOptions;
 import com.google.cloud.bigtable.config.Logger;
 import com.google.cloud.bigtable.grpc.BigtableClusterName;
@@ -34,7 +33,6 @@ import com.google.cloud.bigtable.grpc.BigtableInstanceName;
 import com.google.cloud.bigtable.grpc.BigtableSession;
 import com.google.cloud.bigtable.hbase.BigtableConstants;
 import com.google.cloud.bigtable.hbase.BigtableOptionsFactory;
-import com.google.cloud.bigtable.hbase.adapters.SampledRowKeysAdapter;
 import com.google.cloud.bigtable.hbase.util.ModifyTableBuilder;
 import com.google.cloud.bigtable.hbase2_x.adapters.admin.TableAdapter2x;
 import io.grpc.Status;
@@ -57,7 +55,6 @@ import org.apache.hadoop.hbase.client.CommonConnection;
 import org.apache.hadoop.hbase.client.CompactType;
 import org.apache.hadoop.hbase.client.CompactionState;
 import org.apache.hadoop.hbase.client.RegionInfo;
-import org.apache.hadoop.hbase.client.RegionInfoBuilder;
 import org.apache.hadoop.hbase.client.ServiceCaller;
 import org.apache.hadoop.hbase.client.SnapshotDescription;
 import org.apache.hadoop.hbase.client.TableDescriptor;
@@ -83,6 +80,7 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
+import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.ExecutionException;
 import java.util.function.Function;
 import java.util.function.Predicate;
@@ -105,6 +103,7 @@ public class BigtableAsyncAdmin implements AsyncAdmin {
   private BigtableClusterName bigtableSnapshotClusterName;
   private final Configuration configuration;
   private final BigtableSession bigtableSession;
+  private final CommonConnection commonConnection;
 
   public BigtableAsyncAdmin(CommonConnection commonConnection) throws IOException {
     this.options = commonConnection.getOptions();
@@ -113,6 +112,7 @@ public class BigtableAsyncAdmin implements AsyncAdmin {
     this.bigtableInstanceName = options.getInstanceName();
     this.configuration = commonConnection.getConfiguration();
     this.tableAdapter2x = new TableAdapter2x(options);
+    this.commonConnection = commonConnection;
 
     String clusterId = configuration.get(BigtableOptionsFactory.BIGTABLE_SNAPSHOT_CLUSTER_ID_KEY, null);
     if (clusterId != null) {
@@ -613,31 +613,13 @@ public class BigtableAsyncAdmin implements AsyncAdmin {
 
   @Override
   public CompletableFuture<List<RegionInfo>> getRegions(TableName tableName) {
-    ServerName serverName = ServerName.valueOf(options.getDataHost(), options.getPort(), 0);
-    SampledRowKeysAdapter sampledRowKeysAdapter = getSampledRowKeysAdapter(tableName, serverName);
-    SampleRowKeysRequest.Builder request = SampleRowKeysRequest.newBuilder();
-    request.setTableName(options.getInstanceName().toTableNameStr(tableName.getNameAsString()));
-    return FutureUtils
-        .toCompletableFuture(bigtableSession.getDataClient().sampleRowKeysAsync(request.build()))
-        .thenApplyAsync(result ->
-            sampledRowKeysAdapter.adaptResponse(result)
-                .stream()
-                .map(location -> location.getRegion())
-                .collect(Collectors.toList())
-        );
-  }
-
-  private SampledRowKeysAdapter getSampledRowKeysAdapter(TableName tableNameAdapter,
-      ServerName serverNameAdapter) {
-    return new SampledRowKeysAdapter(tableNameAdapter, serverNameAdapter) {
-      @Override
-      protected HRegionLocation createRegionLocation(byte[] startKey,
-          byte[] endKey) {
-        RegionInfo regionInfo = RegionInfoBuilder.newBuilder(tableNameAdapter)
-            .setStartKey(startKey).setEndKey(endKey).build();
-        return new HRegionLocation(regionInfo, serverNameAdapter);
+    return CompletableFuture.supplyAsync(() -> {
+      try {
+        return new CopyOnWriteArrayList<RegionInfo>(commonConnection.getAllRegionInfos(tableName));
+      } catch (IOException e) {
+        throw new CompletionException(e);
       }
-    };
+    });
   }
 
   private BigtableClusterName getSnapshotClusterName() throws IOException {
