+++ bigtable-hbase-2.x-parent/bigtable-hbase-2.x/src/main/java/com/google/cloud/bigtable/hbase2_x/BigtableAsyncTableRegionLocator.java
@@ -18,20 +18,16 @@ package com.google.cloud.bigtable.hbase2_x;
 import java.util.concurrent.CompletableFuture;
 
 import org.apache.hadoop.hbase.HRegionLocation;
-import org.apache.hadoop.hbase.ServerName;
 import org.apache.hadoop.hbase.TableName;
 import org.apache.hadoop.hbase.client.AsyncTableRegionLocator;
-import org.apache.hadoop.hbase.client.RegionInfo;
-import org.apache.hadoop.hbase.client.RegionInfoBuilder;
 
 import com.google.cloud.bigtable.config.BigtableOptions;
 import com.google.cloud.bigtable.grpc.BigtableDataClient;
 import com.google.cloud.bigtable.hbase.AbstractBigtableRegionLocator;
-import com.google.cloud.bigtable.hbase.adapters.SampledRowKeysAdapter;
 
 /**
  * Bigtable implementation of {@link AsyncTableRegionLocator}
- * 
+ *
  * @author spollapally
  */
 public class BigtableAsyncTableRegionLocator extends AbstractBigtableRegionLocator implements AsyncTableRegionLocator {
@@ -59,18 +55,5 @@ public class BigtableAsyncTableRegionLocator extends AbstractBigtableRegionLocat
           }
           return hRegionLocation;
      });
-  }
-
-  @Override
-  public SampledRowKeysAdapter getSampledRowKeysAdapter(TableName tableName,
-      ServerName serverName) {
-    return new SampledRowKeysAdapter(tableName, serverName) {
-      @Override
-      protected HRegionLocation createRegionLocation(byte[] startKey, byte[] endKey) {
-        RegionInfo regionInfo =
-            RegionInfoBuilder.newBuilder(tableName).setStartKey(startKey).setEndKey(endKey).build();
-        return new HRegionLocation(regionInfo, serverName);
-        }
-      };
   }
 }
