+++ bigtable-hbase-2.x-parent/bigtable-hbase-2.x/src/main/java/org/apache/hadoop/hbase/client/BigtableAsyncConnection.java
@@ -20,20 +20,29 @@ import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
+import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
 
 import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.hbase.HRegionInfo;
+import org.apache.hadoop.hbase.HRegionLocation;
+import org.apache.hadoop.hbase.ServerName;
 import org.apache.hadoop.hbase.TableName;
 import org.apache.hadoop.hbase.security.User;
 
+import com.google.bigtable.v2.SampleRowKeysRequest;
+import com.google.bigtable.v2.SampleRowKeysResponse;
 import com.google.cloud.bigtable.config.BigtableOptions;
 import com.google.cloud.bigtable.config.Logger;
 import com.google.cloud.bigtable.grpc.BigtableSession;
 import com.google.cloud.bigtable.hbase.BigtableOptionsFactory;
 import com.google.cloud.bigtable.hbase.adapters.Adapters;
 import com.google.cloud.bigtable.hbase.adapters.HBaseRequestAdapter;
+import com.google.cloud.bigtable.hbase.adapters.SampledRowKeysAdapter;
 import com.google.cloud.bigtable.hbase.adapters.HBaseRequestAdapter.MutationAdapters;
 import com.google.cloud.bigtable.hbase2_x.BigtableAsyncAdmin;
 import com.google.cloud.bigtable.hbase2_x.BigtableAsyncBufferedMutator;
@@ -272,6 +281,35 @@ public class BigtableAsyncConnection implements AsyncConnection, Closeable, Comm
   @Override
   public AsyncTableRegionLocator getRegionLocator(TableName tableName) {
     return new BigtableAsyncTableRegionLocator(tableName, options, this.session.getDataClient());
+  }
+
+  @Override
+  public List<HRegionInfo> getAllRegionInfos(TableName tableName) throws IOException {
+    ServerName serverName = ServerName.valueOf(options.getDataHost(), options.getPort(), 0);
+    SampledRowKeysAdapter sampledRowKeysAdapter = getSampledRowKeysAdapter(tableName, serverName);
+    SampleRowKeysRequest.Builder request = SampleRowKeysRequest.newBuilder();
+    request.setTableName(options.getInstanceName().toTableNameStr(tableName.getNameAsString()));
+
+    List<SampleRowKeysResponse> result = this.session.getDataClient().sampleRowKeys(request.build());
+
+    return sampledRowKeysAdapter.adaptResponse(result)
+        .stream()
+        .map(location -> location.getRegionInfo())
+        .collect(Collectors.toCollection(CopyOnWriteArrayList::new));
+  }
+
+
+  private SampledRowKeysAdapter getSampledRowKeysAdapter(TableName tableNameAdapter,
+      ServerName serverNameAdapter) {
+    return new SampledRowKeysAdapter(tableNameAdapter, serverNameAdapter) {
+      @Override
+      protected HRegionLocation createRegionLocation(byte[] startKey,
+          byte[] endKey) {
+        RegionInfo regionInfo = RegionInfoBuilder.newBuilder(tableNameAdapter)
+            .setStartKey(startKey).setEndKey(endKey).build();
+        return new HRegionLocation(regionInfo, serverNameAdapter);
+      }
+    };
   }
 
   @Override
